added support for successfull ipc command transmit callback
every command will have a callback to announce that 
it was successfully transmited over ipc -
i.e. its last buffer was successfully sent (its update was received from peer).
This will somehow fix the following situation:

./test "" 1000
./test "" 2000

./alphaclienta 10 localhost 1114 "/home/vpalade/tmp/00000001/" 10 3 localhost 2222

alfter some time, do quit in test 2000.

The client should stop asap.

This is not a complete fix of the above problem. Here is why:

If test 2000 is stopped when sending only responses and test 1000 is only waiting for
responses, test 1000 will not detect the stop of test 2000. 

So the easyest way to fix the problem is for the alpha connections to wait with a 
timeout for receiving the response.

TODO:!!!
Another ideea to improve the time for detecting that a response
will never come, is to implement keep alive at connection level.