some design changes in aioselector and working on sslserver
TODO: review the changes and think if you can simplify the design!!!!

the aioselector changes:
from now on the aioobjects can return OK and have pending requests.
the only difference between OK and NOK is the fact that on OK
the object is added to the execution queue.
this is usefull for objects:
i'm waiting for io completion, but i also have something to do
on reentrance so i exit with OK.